style(checkstyle): change code format following checkstyle rule (#16)
* style: change tab size from 2 to 4

* test(application): remove unused test code

* style(package): remove underscore in package name
