+++ src/main/java/com/e2i/wemeet/WemeetApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class WemeetApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(WemeetApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(WemeetApplication.class, args);
+    }
 
 }
