+++ src/main/java/com/e2i/wemeet/controller/GlobalExceptionController.java
@@ -1,5 +1,7 @@
 package com.e2i.wemeet.controller;
 
+import static com.e2i.wemeet.exception.ErrorCode.UNEXPECTED_INTERNAL;
+
 import com.e2i.wemeet.exception.ErrorResponse;
 import com.e2i.wemeet.exception.badrequest.InvalidValueException;
 import com.e2i.wemeet.exception.internal.InternalServerException;
@@ -12,13 +14,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import static com.e2i.wemeet.exception.ErrorCode.UNEXPECTED_INTERNAL;
-
 @Slf4j
 @RestControllerAdvice
 public class GlobalExceptionController {
 
-    private static final String ERROR_LOG_FORMAT = "Error Class : {}, Error Code : {}, Message : {}";
+    private static final String ERROR_LOG_FORMAT
+        = "Error Class : {}, Error Code : {}, Message : {}";
     private final MessageSourceAccessor messageSourceAccessor;
 
     public GlobalExceptionController(MessageSource messageSource) {
@@ -26,13 +27,14 @@ public class GlobalExceptionController {
     }
 
     @ExceptionHandler(InvalidValueException.class)
-    public ResponseEntity<ErrorResponse> handleInvalidValueException(final InvalidValueException e) {
+    public ResponseEntity<ErrorResponse> handleInvalidValueException(
+        final InvalidValueException e) {
         final int code = e.getErrorCode().getCode();
         final String message = messageSourceAccessor.getMessage(e.getMessage());
 
         log.info(ERROR_LOG_FORMAT, e.getClass().getSimpleName(), code, message);
         return ResponseEntity.ok()
-                .body(new ErrorResponse(code, message));
+            .body(new ErrorResponse(code, message));
     }
 
     @ExceptionHandler(NotFoundException.class)
@@ -42,36 +44,39 @@ public class GlobalExceptionController {
 
         log.info(ERROR_LOG_FORMAT, e.getClass().getSimpleName(), code, message);
         return ResponseEntity.ok()
-                .body(new ErrorResponse(code, message));
+            .body(new ErrorResponse(code, message));
     }
 
     @ExceptionHandler(UnAuthorizedException.class)
-    public ResponseEntity<ErrorResponse> handleUnAuthorizedException(final UnAuthorizedException e) {
+    public ResponseEntity<ErrorResponse> handleUnAuthorizedException(
+        final UnAuthorizedException e) {
         final int code = e.getErrorCode().getCode();
         final String message = messageSourceAccessor.getMessage(e.getMessage());
 
         log.info(ERROR_LOG_FORMAT, e.getClass().getSimpleName(), code, message);
         return ResponseEntity.ok()
-                .body(new ErrorResponse(code, message));
+            .body(new ErrorResponse(code, message));
     }
 
     @ExceptionHandler(InternalServerException.class)
-    public ResponseEntity<ErrorResponse> handleInternalServerException(final InternalServerException e) {
+    public ResponseEntity<ErrorResponse> handleInternalServerException(
+        final InternalServerException e) {
         final int code = e.getErrorCode().getCode();
         final String message = messageSourceAccessor.getMessage(e.getMessage());
 
         log.warn(ERROR_LOG_FORMAT, e.getClass().getSimpleName(), code, message);
         return ResponseEntity.internalServerError()
-                .body(new ErrorResponse(code, message));
+            .body(new ErrorResponse(code, message));
     }
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity<ErrorResponse> handleUnexpectedException(final Exception e) {
         final int code = UNEXPECTED_INTERNAL.getCode();
-        final String message = messageSourceAccessor.getMessage(UNEXPECTED_INTERNAL.getMessageKey());
+        final String message = messageSourceAccessor.getMessage(
+            UNEXPECTED_INTERNAL.getMessageKey());
 
         log.error(ERROR_LOG_FORMAT, e.getClass().getName(), code, message);
         return ResponseEntity.internalServerError()
-                .body(new ErrorResponse(code, message));
+            .body(new ErrorResponse(code, message));
     }
 }
