+++ src/main/java/com/e2i/wemeet/domain/base/BaseTimeEntity.java
@@ -4,21 +4,21 @@ import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
 import java.time.LocalDateTime;
+import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import lombok.Getter;
 
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
 public abstract class BaseTimeEntity {
 
-  @CreatedDate
-  @Column(name = "CREATED_AT")
-  private LocalDateTime createdAt;
+    @CreatedDate
+    @Column(name = "CREATED_AT")
+    private LocalDateTime createdAt;
 
-  @LastModifiedDate
-  @Column(name = "MODIFIED_AT")
-  private LocalDateTime modifiedAt;
+    @LastModifiedDate
+    @Column(name = "MODIFIED_AT")
+    private LocalDateTime modifiedAt;
 }
