+++ src/main/java/com/e2i/wemeet/domain/code/CodePk.java
@@ -2,11 +2,11 @@ package com.e2i.wemeet.domain.code;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
+import java.io.Serializable;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import java.io.Serializable;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
@@ -14,8 +14,8 @@ import java.io.Serializable;
 @Embeddable
 public class CodePk implements Serializable {
 
-  @Column(length = 4, name = "codeId")
-  private String codeId;
+    @Column(length = 4, name = "codeId")
+    private String codeId;
 
-  private String groupCodeId;
+    private String groupCodeId;
 }
