+++ src/main/java/com/e2i/wemeet/domain/code/GroupCode.java
@@ -3,8 +3,8 @@ package com.e2i.wemeet.domain.code;
 import com.e2i.wemeet.domain.base.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,13 +15,13 @@ import lombok.NoArgsConstructor;
 @Entity
 public class GroupCode extends BaseTimeEntity {
 
-  @Id
-  @Column(length = 4)
-  private String groupCodeId;
+    @Id
+    @Column(length = 4)
+    private String groupCodeId;
 
-  @Column(nullable = false)
-  private String groupCodeName;
+    @Column(nullable = false)
+    private String groupCodeName;
 
-  @Column(nullable = false)
-  private String description;
+    @Column(nullable = false)
+    private String description;
 }
