+++ src/main/java/com/e2i/wemeet/domain/member/CollegeInfo.java
@@ -12,23 +12,23 @@ import lombok.NoArgsConstructor;
 @Embeddable
 public class CollegeInfo {
 
-  @Column(length = 30, nullable = false)
-  private String college;
+    @Column(length = 30, nullable = false)
+    private String college;
 
-  @Column(length = 20, nullable = false)
-  private String collegeType;
+    @Column(length = 20, nullable = false)
+    private String collegeType;
 
-  @Column(nullable = false)
-  private int admissionYear;
+    @Column(nullable = false)
+    private int admissionYear;
 
-  @Column(length = 50, unique = true)
-  private String mail;
+    @Column(length = 50, unique = true)
+    private String mail;
 
-  @Builder
-  public CollegeInfo(String college, String collegeType, int admissionYear, String mail) {
-    this.college = college;
-    this.collegeType = collegeType;
-    this.admissionYear = admissionYear;
-    this.mail = mail;
-  }
+    @Builder
+    public CollegeInfo(String college, String collegeType, int admissionYear, String mail) {
+        this.college = college;
+        this.collegeType = collegeType;
+        this.admissionYear = admissionYear;
+        this.mail = mail;
+    }
 }
