+++ src/main/java/com/e2i/wemeet/domain/member/Mbti.java
@@ -7,16 +7,16 @@ import java.util.Arrays;
 
 public enum Mbti {
 
-  ESTJ, ESTP, ESFJ, ESFP, ENTJ, ENTP, ENFJ, ENFP,
-  ISTJ, ISTP, ISFJ, ISFP, INTJ, INTP, INFJ, INFP,
-  NOTHING;
+    ESTJ, ESTP, ESFJ, ESFP, ENTJ, ENTP, ENFJ, ENFP,
+    ISTJ, ISTP, ISFJ, ISFP, INTJ, INTP, INFJ, INFP,
+    NOTHING;
 
-  @JsonCreator
-  public static Mbti findBy(String value) {
+    @JsonCreator
+    public static Mbti findBy(String value) {
 
-    return Arrays.stream(Mbti.values())
-        .filter(mbti -> mbti.name().equals(value.toUpperCase()))
-        .findFirst()
-        .orElseThrow(() -> new InvalidValueException(ErrorCode.INVALID_MBTI_VALUE));
-  }
+        return Arrays.stream(Mbti.values())
+            .filter(mbti -> mbti.name().equals(value.toUpperCase()))
+            .findFirst()
+            .orElseThrow(() -> new InvalidValueException(ErrorCode.INVALID_MBTI_VALUE));
+    }
 }
