+++ src/main/java/com/e2i/wemeet/domain/member/Member.java
@@ -25,57 +25,58 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Member extends BaseTimeEntity {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long memberId;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long memberId;
 
-  @Column(length = 4, nullable = false)
-  private String memberCode;
+    @Column(length = 4, nullable = false)
+    private String memberCode;
 
-  @Column(length = 20, nullable = false)
-  private String nickname;
+    @Column(length = 20, nullable = false)
+    private String nickname;
 
-  @Column(length = 6, nullable = false)
-  @Enumerated(value = EnumType.STRING)
-  private Gender gender;
+    @Column(length = 6, nullable = false)
+    @Enumerated(value = EnumType.STRING)
+    private Gender gender;
 
-  @Column(length = 13, unique = true, nullable = false)
-  private String phoneNumber;
+    @Column(length = 13, unique = true, nullable = false)
+    private String phoneNumber;
 
-  @Embedded
-  private CollegeInfo collegeInfo;
+    @Embedded
+    private CollegeInfo collegeInfo;
 
-  @Embedded
-  private Preference preference;
+    @Embedded
+    private Preference preference;
 
-  @Column(length = 7, nullable = false)
-  @Enumerated(value = EnumType.STRING)
-  private Mbti mbti;
+    @Column(length = 7, nullable = false)
+    @Enumerated(value = EnumType.STRING)
+    private Mbti mbti;
 
-  @Column(length = 100)
-  private String introduction;
+    @Column(length = 100)
+    private String introduction;
 
-  @Column(nullable = false)
-  private int credit;
+    @Column(nullable = false)
+    private int credit;
 
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "teamId")
-  private Team team;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "teamId")
+    private Team team;
 
-  @Builder
-  public Member(Long memberId, String memberCode, String nickname, Gender gender,
-      String phoneNumber, CollegeInfo collegeInfo, Preference preference, Mbti mbti,
-      String introduction, int credit, Team team) {
-    this.memberId = memberId;
-    this.memberCode = memberCode;
-    this.nickname = nickname;
-    this.gender = gender;
-    this.phoneNumber = phoneNumber;
-    this.collegeInfo = collegeInfo;
-    this.preference = preference;
-    this.mbti = mbti;
-    this.introduction = introduction;
-    this.credit = credit;
-    this.team = team;
-  }
+    @Builder
+    public Member(Long memberId, String memberCode, String nickname, Gender gender,
+        String phoneNumber, CollegeInfo collegeInfo, Preference preference, Mbti mbti,
+        String introduction, int credit, Team team) {
+        this.memberId = memberId;
+        this.memberCode = memberCode;
+        this.nickname = nickname;
+        this.gender = gender;
+        this.phoneNumber = phoneNumber;
+        this.collegeInfo = collegeInfo;
+        this.preference = preference;
+        this.mbti = mbti;
+
+        this.introduction = introduction;
+        this.credit = credit;
+        this.team = team;
+    }
 }
