--- src/main/java/com/e2i/wemeet/domain/member_preference_meeting_type/MemberPreferenceMeetingTypeRepository.java
@@ -1,7 +1,0 @@
-package com.e2i.wemeet.domain.member_preference_meeting_type;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberPreferenceMeetingTypeRepository extends JpaRepository<MemberPreferenceMeetingType, Long> {
-
-}
