+++ src/main/java/com/e2i/wemeet/domain/memberinterest/MemberInterestRepository.java
@@ -1,4 +1,4 @@
-package com.e2i.wemeet.domain.member_interest;
+package com.e2i.wemeet.domain.memberinterest;
 
 import org.springframework.data.jpa.repository.JpaRepository;
