+++ src/main/java/com/e2i/wemeet/domain/memberpreferencemeetingtype/MemberPreferenceMeetingTypeRepository.java
@@ -1,0 +1,8 @@
+package com.e2i.wemeet.domain.memberpreferencemeetingtype;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberPreferenceMeetingTypeRepository extends
+    JpaRepository<MemberPreferenceMeetingType, Long> {
+
+}
