+++ src/main/java/com/e2i/wemeet/domain/team/Team.java
@@ -25,35 +25,35 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Team extends BaseTimeEntity {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long teamId;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long teamId;
 
-  @Column(nullable = false)
-  private int memberCount;
+    @Column(nullable = false)
+    private int memberCount;
 
-  @Column(length = 6, nullable = false)
-  @Enumerated(value = EnumType.STRING)
-  private Gender gender;
+    @Column(length = 6, nullable = false)
+    @Enumerated(value = EnumType.STRING)
+    private Gender gender;
 
-  @Column(nullable = false)
-  private boolean drinkingOption;
+    @Column(nullable = false)
+    private boolean drinkingOption;
 
-  @Column(length = 100)
-  private String introduction;
+    @Column(length = 100)
+    private String introduction;
 
-  @OneToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "memberId")
-  private Member member;
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "memberId")
+    private Member member;
 
-  @Builder
-  public Team(Long teamId, int memberCount, Gender gender, boolean drinkingOption,
-      String introduction, Member member) {
-    this.teamId = teamId;
-    this.memberCount = memberCount;
-    this.gender = gender;
-    this.drinkingOption = drinkingOption;
-    this.introduction = introduction;
-    this.member = member;
-  }
+    @Builder
+    public Team(Long teamId, int memberCount, Gender gender, boolean drinkingOption,
+        String introduction, Member member) {
+        this.teamId = teamId;
+        this.memberCount = memberCount;
+        this.gender = gender;
+        this.drinkingOption = drinkingOption;
+        this.introduction = introduction;
+        this.member = member;
+    }
 }
