+++ src/main/java/com/e2i/wemeet/exception/CustomException.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class CustomException extends RuntimeException {
+
     private final ErrorCode errorCode;
 
     public CustomException(final ErrorCode errorCode) {
