+++ src/main/java/com/e2i/wemeet/exception/ErrorCode.java
@@ -7,8 +7,7 @@ public enum ErrorCode {
     INVALID_SMS_CREDENTIAL(40001, "invalid.sms.credential"),
     INVALID_GENDER_VALUE(40002, "invalid.gender.value"),
     INVALID_MBTI_VALUE(40003, "invalid.mbti.value"),
-    UNEXPECTED_INTERNAL(50000, "unexpected.internal")
-    ;
+    UNEXPECTED_INTERNAL(50000, "unexpected.internal");
 
     private final int code;
     private final String messageKey;
