+++ src/main/java/com/e2i/wemeet/exception/ErrorResponse.java
@@ -1,4 +1,5 @@
 package com.e2i.wemeet.exception;
 
 public record ErrorResponse(int code, String message) {
+
 }
