+++ src/main/java/com/e2i/wemeet/exception/badrequest/InvalidSmsCredentialException.java
@@ -1,12 +1,12 @@
 package com.e2i.wemeet.exception.badrequest;
 
-import com.e2i.wemeet.exception.ErrorCode;
-
 import static com.e2i.wemeet.exception.ErrorCode.INVALID_SMS_CREDENTIAL;
 
+import com.e2i.wemeet.exception.ErrorCode;
+
 /*
-* SMS 문자 인증 실패시 발생
-*/
+ * SMS 문자 인증 실패시 발생
+ */
 public class InvalidSmsCredentialException extends InvalidValueException {
 
     public InvalidSmsCredentialException() {
