--- src/test/java/com/e2i/wemeet/WemeetApplicationTests.java
@@ -1,13 +1,0 @@
-package com.e2i.wemeet;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WemeetApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
